backend/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentification JWT
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-decouple==3.8

# Base de données
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Ollama et AI
ollama==0.1.8
httpx>=0.27.0
numpy>=1.24.0
tiktoken>=0.5.0
sentence-transformers

# Support Excel et données
pandas>=2.0.0
openpyxl>=3.1.0

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilitaires
python-dotenv==1.0.0
loguru==0.7.2
typing-extensions==4.8.0
aiofiles==23.2.1
websockets==12.0

# Supprimer les dépendances de développement inutiles : pytest, pytest-asyncio, black, flake8
# CORS est intégré dans FastAPI, pas besoin de package séparé